Fixed reliability issues

+++ src/main/java/fr/aerwyn81/headblocks/utils/message/color/IridiumColorAPI.java
@@ -251,7 +251,7 @@ public class IridiumColorAPI {
         double nearestDistance = Integer.MAX_VALUE;
 
         for (Color constantColor : COLORS.keySet()) {
-            double distance = Math.pow(color.getRed() - constantColor.getRed(), 2) + Math.pow(color.getGreen() - constantColor.getGreen(), 2) + Math.pow(color.getBlue() - constantColor.getBlue(), 2);
+            double distance = Math.pow((double) color.getRed() - constantColor.getRed(), 2) + Math.pow((double) color.getGreen() - constantColor.getGreen(), 2) + Math.pow((double) color.getBlue() - constantColor.getBlue(), 2);
             if (nearestDistance > distance) {
                 nearestColor = constantColor;
                 nearestDistance = distance;
